refactor: reorganize project structure by moving context, hooks, and types to application layer
