+++ src/presentation/widgets/modals/CreateProjectModal.tsx
@@ -1,8 +1,8 @@
 import { BaseModal } from "./BaseModal";
 import { BaseModalProps } from "./modal.types";
 import { CreateProjectForm } from "@/presentation/widgets/projects/CreateProjectForm";
-import { CreateProjectFormData } from "@/presentation/widgets/projects/types";
-import { useProjectCreation } from "@/presentation/widgets/projects/context/ProjectCreationContext";
+import { CreateProjectFormData } from "@/application/projects/types";
+import { useProjectCreation } from "@/application/projects/context/ProjectCreationContext";
 
 export default function CreateProjectModal(props: BaseModalProps) {
   const { isOpen, onClose, onBack } = props;
