+++ src/presentation/widgets/projects/CreateProjectForm/CreateProjectForm.types.ts
@@ -1,4 +1,4 @@
-import { CreateProjectFormData } from "../types";
+import { CreateProjectFormData } from "../../../../application/projects/types";
 
 export type { CreateProjectFormData };
 export type FormValues = CreateProjectFormData;
